requirements.txt
absl-py==1.1.0
albumentations @ git+https://github.com/albumentations-team/albumentations@89a675cbfb2b76f6be90e7049cd5211cb08169a5
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
attrs==21.4.0
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
beautifulsoup4==4.11.1
bleach==5.0.1
cachetools==5.2.0
certifi==2022.5.18.1
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
cycler==0.11.0
dataclasses==0.6
debugpy @ file:///tmp/build/80754af9/debugpy_1637091426235/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
defusedxml==0.7.1
docker-pycreds==0.4.0
efficientnet-pytorch==0.6.3
entrypoints @ file:///tmp/build/80754af9/entrypoints_1649908585034/work
fastjsonschema==2.15.3
fonttools==4.33.3
future==0.18.2
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.9.0
google-auth-oauthlib==0.4.6
grpcio==1.47.0
idna==3.3
imageio==2.19.3
importlib-metadata==4.12.0
importlib-resources==5.8.0
ipykernel @ file:///tmp/build/80754af9/ipykernel_1646983213839/work/dist/ipykernel-6.9.1-py3-none-any.whl
ipython @ file:///tmp/build/80754af9/ipython_1643818144432/work
ipython-genutils==0.2.0
ipywidgets==7.7.1
jedi @ file:///tmp/build/80754af9/jedi_1644299024593/work
Jinja2==3.1.2
joblib==1.1.0
jsonschema==4.7.2
jupyter==1.0.0
jupyter-client @ file:///opt/conda/conda-bld/jupyter_client_1650622202839/work
jupyter-console==6.4.4
jupyter-core @ file:///opt/conda/conda-bld/jupyter_core_1651671229925/work
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.1
kiwisolver==1.4.2
Markdown==3.3.7
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline @ file:///tmp/build/80754af9/matplotlib-inline_1628242447089/work
mistune==0.8.4
munch==2.5.0
nbclient==0.6.6
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio @ file:///tmp/build/80754af9/nest-asyncio_1649847911453/work
networkx==2.6.3
notebook==6.4.12
numpy==1.21.6
numpyencoder==0.3.0
oauthlib==3.2.0
onnx==1.12.0
opencv-python==4.5.5.64
opencv-python-headless==4.5.5.64
packaging==21.3
pandas==1.3.5
pandocfilters==1.5.0
parse==1.19.0
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
pathtools==0.1.2
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==9.1.1
pretrainedmodels==0.7.4
prometheus-client==0.14.1
promise==2.3
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1633440160888/work
protobuf==3.19.4
psutil==5.9.1
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments @ file:///opt/conda/conda-bld/pygments_1644249106324/work
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz==2022.1
PyWavelets==1.3.0
PyYAML==6.0
pyzmq @ file:///tmp/build/80754af9/pyzmq_1638434924971/work
qtconsole==5.3.1
QtPy==2.1.0
qudida==0.0.4
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-image==0.18.3
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.11.2
segmentation-models-pytorch==0.2.1
Send2Trash==1.8.0
sentry-sdk==1.7.0
setproctitle==1.2.3
shortuuid==1.0.9
six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap==5.0.0
soupsieve==2.3.2.post1
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
terminado==0.15.0
thop==0.1.0.post2206102148
threadpoolctl==3.1.0
tifffile==2021.11.2
timm==0.4.12
tinycss2==1.1.1
torch==1.7.0+cu110
torchaudio==0.7.0
torchvision==0.8.1+cu110
tornado @ file:///tmp/build/80754af9/tornado_1606942283357/work
tqdm==4.64.0
traitlets==5.3.0
typing_extensions==4.2.0
urllib3==1.26.9
wandb==0.12.21
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
webencodings==0.5.1
Werkzeug==2.1.2
widgetsnbextension==3.6.1
zipp==3.8.0
